fix for tracker-[coppa-Defects][24230] Unable to update Drug Information –Throws HTTP Status 404 Exception - missing field error tag
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@7639 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ pa/pa-ejb/src/java/gov/nih/nci/pa/util/PAUtil.java
@@ -879,7 +879,7 @@ public class PAUtil {
     public static BigDecimal convertStringToDecimal(String value) {
         BigDecimal bd = null;
         if (value != null && isNumber(value)) {
-            bd = BigDecimal.valueOf(Long.parseLong(value));
+            bd = BigDecimal.valueOf(Double.parseDouble(value));
         }
         return bd;
     }
@@ -894,7 +894,7 @@ public class PAUtil {
     public static boolean isNumber(String value) {
       boolean isNumber = true;
       try {
-         Long.parseLong(value);
+         Double.parseDouble(value);
       } catch (NumberFormatException e) {
          isNumber = false;
       }
